fix bug login, refactor code update profile
